+++ indico/modules/rb/client/js/modules/bookRoom/BookRoomModal.jsx
@@ -629,6 +629,28 @@ class BookRoomModal extends React.Component {
                     fprops.form.mutators
                   )}
               </Form>
+
+              {room.canUserViewInternalNotes && (
+                <Form
+                  id="book-notes-form"
+                  onSubmit={fprops.handleSubmit}
+                  style={{paddingTop: '1em'}}
+                >
+                  <Segment inverted styleName="internal-notes-segment">
+                    <h3 style={{marginBottom: '0.5em'}}>
+                      <Icon name="clipboard list" />
+                      <Translate>Internal notes (optional)</Translate>
+                    </h3>
+                    <p>
+                      <Translate>
+                        Internal notes about the booking are only visible to room managers.
+                      </Translate>
+                    </p>
+                    <FinalTextArea name="internalNote" disabled={fprops.submitSucceeded} />
+                  </Segment>
+                </Form>
+              )}
+
               {preConflictsExist && !fprops.submitSucceeded && this.renderPreConflictMessage()}
               {conflictsExist &&
                 this.renderBookingConstraints(
@@ -719,7 +741,7 @@ export default connect(
         fetchRelatedEvents: actions.fetchRelatedEvents,
         resetRelatedEvents: actions.resetRelatedEvents,
         createBooking: (data, props) => {
-          const {reason, usage, user, linkType, linkId, linkBack} = data;
+          const {reason, internalNote, usage, user, linkType, linkId, linkBack} = data;
           const {
             bookingData: {recurrence, dates, timeSlot, isPrebooking},
             room,
@@ -728,6 +750,7 @@ export default connect(
           return actions.createBooking(
             {
               reason,
+              internalNote,
               usage,
               user,
               recurrence,